:technologist: add devcontainer
See https://docs.github.com/en/codespaces/setting-up-your-project-for-codespaces/adding-a-dev-container-configuration/introduction-to-dev-containers

Since there does not seem to be a container image that contains both PowerShell and Git, the PowerShell image is used and Git is installed as a feature. Setup time is shorter compared to a container with Git where PowerShell needs to be installed.